Load datatable asynchronously
Adding all proteins into the model turned turned out
to be too heavy with large amounts of protein. The combination
of injected into the model and having to load the DOM with all
rows before pagination kicks in slowed everything to a crawl.

We now no longer inject it into the model and load it via ajax,
generating the table rows on the fly.

This can be further improved by only loading the required page
but this will require some improvements to the backend and is
not strictly necessary at the moment.